Adding list of end-to-end tests to ignore
The CI gate has been configured to check the file .e2e.test.ignore
for a list of end-to-end tests to ignore.

This commit add a sample of the file, including one conformance
test that is failing, but that seems not related to Midonet
Kubernetes integration. Despite being ignored now, it should be
checked in the future to include it again, as it is required by
Kubernetes conformance approval.

Ref: MINOR_CHANGE

Change-Id: Ieb423cfe0512ce0b338663853ba49748831b0ced
Signed-off-by: Miguel Herranz <miguel@midokura.com>