pod: Recognize mac address annotation
If the mac address is known, create the correspoinding
MAC Port resource.  This is merely an optimization.

Probably we can make the CNI add this annotation
automatically.  To do that, we need to give k8s api
credential to the CNI.
Alternatively, we can somehow generate the MAC address
deterministically.  My impression is 24 bits are too
small to assume no conflicts even if it needs to be
unique within the node though.
Or we can generate the MAC address from the IP address,
assuming that the podCIDR subnet is not huge.

Ref: MNA-1232
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I8538f3192ab7ff4deb87a37ab159dcd9b71e7556