A script to generate Dockefiles from templates
As these files are getting more complex, it's more of burdern
to keep 5 copies of them in sync manually.

build.sh now checks if those Dockerfiles are up-to-date before
building images.

It's still developers' responsibility to re-generate these files
and commit them to the repo.  It's done this way because Docker Hub
automated build doesn't seem to allow to generate Dockerfiles
on demand.

Extra blank linkes added to Dockefiles are side effects of
m4 include.

Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: If9a962ae059c3218f1a93e5b8645cf7541dc678e